backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-firestore
pytest
httpx
google-generativeai
python-dotenv
firebase-admin
